feat: Update getUrlsAltPages method in htmlget.py
This commit updates the `getUrlsAltPages` method in `htmlget.py` to include additional pages in the list of URLs to scrape. It calculates the number of pages based on the total number of products and adds the corresponding URLs to the list. If an error occurs while processing a page, an error message is printed.